+++ sdk/dynamicBrcode/index.js
@@ -1,0 +1,10 @@
+const dynamicBrcode = require('./dynamicBrcode.js');
+
+exports.create = dynamicBrcode.create;
+exports.query = dynamicBrcode.query;
+exports.get = dynamicBrcode.get;
+exports.page = dynamicBrcode.page;
+exports.verify = dynamicBrcode.verify;
+exports.responseInstant = dynamicBrcode.responseInstant;
+exports.responseDue = dynamicBrcode.responseDue;
+exports.DynamicBrcode = dynamicBrcode.DynamicBrcode;