+++ sdk/issuingInvoice/issuingInvoice.js
@@ -16,7 +16,7 @@ class IssuingInvoice extends Resource {
      * Parameters (optional):
      * @param taxId [string, default null]: payer tax ID (CPF or CNPJ) with or without formatting. ex: '01234567890' or '20.018.183/0001-80'
      * @param name [string, default null]: payer name. ex: 'Iron Bank S.A.'
-     * @param tags [array of strings, default []]: array of strings for tagging
+     * @param tags [list of strings, default []]: list of strings for tagging
      *
      * Attributes (return-only):
      * @param id [string]: unique id returned when the IssuingInvoice is created. ex: '5656565656565656'
@@ -50,7 +50,7 @@ exports.create = async function ({ amount, name=null, taxId=null, tags=null, use
      * @description Send a list of IssuingInvoice objects for creation in the Stark Infra API
      *
      * Parameters (required):
-     * @param invoices [array of IssuingInvoice objects]: array of IssuingInvoice objects to be created in the API
+     * @param invoices [list of IssuingInvoice objects]: list of IssuingInvoice objects to be created in the API
      *
      * Parameters (optional):
      * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkinfra.user was set before function call
@@ -101,11 +101,11 @@ exports.query = async function ({ status, after, before, tags, limit, user } = {
      * @param after [DateTime or string, default null] date filter for objects created only after specified date. ex: '2020-04-03'
      * @param before [DateTime or string, default null] date filter for objects created only before specified date. ex: '2020-04-03'
      * @param status [string, default null]: filter for status of retrieved objects. ex: 'created', 'paid', 'canceled' or 'overdue'
-     * @param tags [array of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
+     * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
-     * @returns generator of IssuingInvoices objects with updated attributes
+     * @returns generator of IssuingInvoice objects with updated attributes
      *
      */
     let query = {
@@ -132,7 +132,7 @@ exports.page = async function ({ cursor, status, after, before, tags, limit, use
      * @param after [string, default null] date filter for objects created only after specified date. ex: '2020-04-03'
      * @param before [string, default null] date filter for objects created only before specified date. ex: '2020-04-03'
      * @param status [string, default null]: filter for status of retrieved objects. ex: 'created', 'paid', 'canceled' or 'overdue'
-     * @param tags [array of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
+     * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return: