+++ sdk/merchantCategory/index.js
@@ -1,0 +1,4 @@
+const merchantCategory = require('./merchantCategory.js');
+
+exports.query = merchantCategory.query;
+exports.MerchantCategory = merchantCategory.MerchantCategory;