+++ sdk/pixReversal/pixReversal.js
@@ -20,12 +20,11 @@ class PixReversal extends Resource {
      * @param reason [string]: reason why the PixReversal is being reversed. Options are 'bankError', 'fraud', 'chashierError', 'customerRequest'
      *
      * Parameters (optional):
-     * @param tags [array of strings, default null]: list of strings for reference when searching for PixReversals. ex: ['employees', 'monthly']
+     * @param tags [list of strings, default null]: list of strings for reference when searching for PixReversals. ex: ['employees', 'monthly']
      *
      * Attributes (return-only):
      * @param id [string]: unique id returned when the PixReversal is created. ex: '5656565656565656'
      * @param returnId [string]: central bank's unique reversal transaction ID. ex: 'D20018183202202030109X3OoBHG74wo'.
-     * @param bankCode [string]: code of the bank institution in Brazil. ex: '20018183' or '341'
      * @param fee [integer]: fee charged when PixReversal is paid. ex: 200 (= R$ 2.00)
      * @param status [string]: current PixReversal status. ex: 'registered' or 'paid'
      * @param flow [string]: direction of money flow. ex: 'in' or 'out'
@@ -33,8 +32,8 @@ class PixReversal extends Resource {
      * @param updated [string]: latest update datetime for the PixReversal. ex: '2020-03-10 10:30:00.000'
      *
      */
-    constructor({ amount, externalId, endToEndId, reason, tags, id, returnId, bankCode, fee, status,
-                    flow, created, updated}) {
+    constructor({ amount, externalId, endToEndId, reason, tags=null, id=null, returnId=null, fee=null, status=null,
+                    flow=null, created=null, updated=null}) {
         super(id);
         this.amount = amount;
         this.externalId = externalId;
@@ -43,7 +42,6 @@ class PixReversal extends Resource {
         this.tags = tags;
         this.id = id;
         this.returnId = returnId;
-        this.bankCode = bankCode;
         this.fee = fee;
         this.status = status;
         this.flow = flow;
@@ -106,11 +104,11 @@ exports.query = async function ({ limit, after, before, status, tags, ids, retur
      * @param limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
      * @param after [string, default null]: date filter for objects created or updated only after specified date. ex: '2020-03-10'
      * @param before [string, default null]: date filter for objects created or updated only before specified date. ex: '2020-03-10'
-     * @param status [array of strings, default null]: filter for status of retrieved objects. ex: 'success' or 'failed'
-     * @param tags [array of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
-     * @param ids [array of strings, default null]: list of ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
-     * @param returnIds [array of strings, default null]: central bank's unique transaction IDs. ex: ['E79457883202101262140HHX553UPqeq', 'E79457883202101262140HHX553UPxzx']
-     * @param externalIds [array of strings, default null]: url safe strings that must be unique among all your PixReversals. Duplicated external IDs will cause failures. By default, this parameter will block any PixReversals that repeats amount and receiver information on the same date. ex: ['my-internal-id-123456', 'my-internal-id-654321']
+     * @param status [list of strings, default null]: filter for status of retrieved objects. ex: 'success' or 'failed'
+     * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
+     * @param ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
+     * @param returnIds [list of strings, default null]: central bank's unique transaction IDs. ex: ['E79457883202101262140HHX553UPqeq', 'E79457883202101262140HHX553UPxzx']
+     * @param externalIds [list of strings, default null]: url safe strings that must be unique among all your PixReversals. Duplicated external IDs will cause failures. By default, this parameter will block any PixReversals that repeats amount and receiver information on the same date. ex: ['my-internal-id-123456', 'my-internal-id-654321']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
@@ -143,11 +141,11 @@ exports.page = async function ({ cursor, limit, after, before, status, tags, ids
      * @param limit [integer, default 100]: maximum number of objects to be retrieved. It must be an integer between 1 and 100. ex: 35
      * @param after [string, default null]: date filter for objects created or updated only after specified date. ex: '2020-03-10'
      * @param before [string, default null]: date filter for objects created or updated only before specified date. ex: '2020-03-10'
-     * @param status [array of strings, default null]: filter for status of retrieved objects. ex: 'success' or 'failed'
-     * @param tags [array of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
-     * @param ids [array of strings, default null]: list of ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
-     * @param endToEndIds [array of strings, default null]: central bank's unique transaction IDs. ex: ['E79457883202101262140HHX553UPqeq', 'E79457883202101262140HHX553UPxzx']
-     * @param externalIds [array of strings, default null]: url safe strings that must be unique among all your PixReversals. Duplicated external IDs will cause failures. By default, this parameter will block any PixReversals that repeats amount and receiver information on the same date. ex: ['my-internal-id-123456', 'my-internal-id-654321']
+     * @param status [list of strings, default null]: filter for status of retrieved objects. ex: 'success' or 'failed'
+     * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
+     * @param ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']
+     * @param returnIds [list of strings, default null]: central bank's unique transaction IDs. ex: ['E79457883202101262140HHX553UPqeq', 'E79457883202101262140HHX553UPxzx']
+     * @param externalIds [list of strings, default null]: url safe strings that must be unique among all your PixReversals. Duplicated external IDs will cause failures. By default, this parameter will block any PixReversals that repeats amount and receiver information on the same date. ex: ['my-internal-id-123456', 'my-internal-id-654321']
      * @param user [Organization/Project object, default null]: Project object. Not necessary if starkinfra.user was set before function call
      *
      * Return:
@@ -171,7 +169,7 @@ exports.page = async function ({ cursor, limit, after, before, status, tags, ids
 exports.parse = async function ({content, signature, user} = {}) {
     /**
      *
-     * Create single verified PixReversal object from a content string
+     * Create a single verified PixReversal object from a content string
      *
      * @description Create a single PixReversal object from a content string received from a handler listening at
      * the reversal url. If the provided digital signature does not check out with the Stark public key, a
@@ -189,4 +187,25 @@ exports.parse = async function ({content, signature, user} = {}) {
      *
      */
     return parse.parseAndVerify(resource, content, signature, user);
+};
+
+exports.response = async function ({status, reason} = {}) {
+    /**
+     *
+     * Helps you respond to a PixReversal authorization
+     *
+     * Parameters (required):
+     * @param status [string]: response to the authorization. Options: 'approved' or 'denied'
+     *
+     * Parameters (conditionally required):
+     * @param reason [string, default null]: denial reason. Options: 'invalidAccountNumber', 'blockedAccount', 'accountClosed', 'invalidAccountType', 'invalidTransactionType', 'taxIdMismatch', 'invalidTaxId', 'orderRejected', 'reversalTimeExpired', 'settlementFailed'
+     *
+     * Return:
+     * @returns Dumped JSON string that must be returned to us
+     *
+     */
+    return JSON.stringify({
+        'status': status,
+        'reason': reason,
+    });
 };